requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.3
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
apache-airflow==3.1.7
apache-airflow-core==3.1.7
apache-airflow-providers-celery==3.15.2
apache-airflow-providers-common-compat==1.13.0
apache-airflow-providers-common-io==1.7.1
apache-airflow-providers-common-sql==1.30.4
apache-airflow-providers-smtp==2.4.2
apache-airflow-providers-standard==1.11.0
apache-airflow-task-sdk==1.1.7
argcomplete==3.6.3
asgiref==3.11.0
attrs==25.4.0
babel==2.18.0
billiard==4.2.4
cadwyn==5.4.6
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorlog==6.10.1
cron-descriptor==2.0.6
croniter==6.0.0
cryptography==42.0.8
deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.20
flower==2.0.1
fsspec==2026.1.0
googleapis-common-protos==1.72.0
greenback==1.3.0
greenlet==3.3.1
grpcio==1.65.5
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanize==4.15.0
idna==3.11
importlib-metadata==8.4.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.0.3
lockfile==0.12.2
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.20.0
natsort==8.4.0
numpy==2.4.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.0
pathspec==1.0.4
pendulum==3.2.0
pluggy==1.6.0
prometheus-client==0.24.1
prompt-toolkit==3.0.52
protobuf==4.25.8
psutil==7.2.2
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pygtrie==2.5.0
pyjwt==2.11.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.3
redis==6.4.0
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rich-argparse==1.7.2
rich-toolkit==0.18.1
rpds-py==0.30.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.46
sqlalchemy-jsonfield==1.0.2
sqlalchemy-utils==0.42.1
sqlparse==0.5.5
starlette==0.48.0
structlog==25.5.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.3.0
text-unidecode==1.3
tornado==6.5.4
typer==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
uc-micro-py==1.0.3
universal-pathlib==0.2.6
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.5.3
websockets==15.0.1
wirerope==1.0.0
wrapt==2.1.1
zipp==3.23.0
